Add public feed generation and improve article parsing
Added static site output including Atom feed, index, and state files under public/. Improved article parsing to handle tag_list as string or list and reactions as i64. Updated HTTP client to set user agent and accept headers. Fixed filtering logic to support i64 reaction counts.

+++ src/main.rs
@@ -43,7 +43,7 @@ fn run() -> Result<(), AppError> {
     let articles = fetch_articles(&config)?;
     let filtered: Vec<_> = articles
         .into_iter()
-        .filter(|item| item.public_reactions_count.unwrap_or(0) >= config.min_reactions)
+        .filter(|item| item.public_reactions_count.unwrap_or(0) >= config.min_reactions as i64)
         .collect();
 
     let merged = state.merge_from_api(&filtered, now);
